+++ cpp/test/daynaport_test.cpp
@@ -84,7 +84,7 @@ TEST_F(DaynaportTest, TestUnitReady)
 
 TEST_F(DaynaportTest, WriteData)
 {
-    array<int, 6> cdb = { };
+    array<uint8_t, 6> cdb = { };
     const array<const uint8_t, 5> buf = { };
 
     cdb[0] = to_underlying(SEND_MESSAGE_6);
@@ -115,27 +115,23 @@ TEST_F(DaynaportTest, GetMessage6)
 
     controller->SetCdbByte(4, 0x00);
     controller->SetCdbByte(5, 0xff);
-    Dispatch(daynaport, GET_MESSAGE_6, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB,
-        "Invalid data format");
+    Dispatch(daynaport, GET_MESSAGE_6, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB, "Invalid data format");
 }
 
 TEST_F(DaynaportTest, SendMessage6)
 {
     controller->SetCdbByte(5, 0x00);
-    Dispatch(daynaport, SEND_MESSAGE_6, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB,
-        "Invalid transfer length");
+    Dispatch(daynaport, SEND_MESSAGE_6, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB, "Invalid transfer length");
 
     controller->SetCdbByte(3, 255);
     controller->SetCdbByte(4, 0);
     controller->SetCdbByte(5, 0x08);
-    Dispatch(daynaport, SEND_MESSAGE_6, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB,
-        "Invalid transfer length");
+    Dispatch(daynaport, SEND_MESSAGE_6, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB, "Invalid transfer length");
 
     controller->SetCdbByte(3, 0);
     controller->SetCdbByte(4, 0);
     controller->SetCdbByte(5, 0xff);
-    Dispatch(daynaport, SEND_MESSAGE_6, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB,
-        "Invalid transfer length");
+    Dispatch(daynaport, SEND_MESSAGE_6, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB, "Invalid transfer length");
 
     controller->SetCdbByte(5, 0x80);
     EXPECT_CALL(*controller, DataOut);
@@ -152,8 +148,7 @@ TEST_F(DaynaportTest, TestRetrieveStats)
 
 TEST_F(DaynaportTest, SetInterfaceMode)
 {
-    Dispatch(daynaport, SET_IFACE_MODE, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB,
-        "Unknown interface command");
+    Dispatch(daynaport, SET_IFACE_MODE, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB, "Unknown interface command");
 
     // Not implemented, do nothing
     controller->SetCdbByte(5, DaynaPort::CMD_SCSILINK_SETMODE);
@@ -166,22 +161,18 @@ TEST_F(DaynaportTest, SetInterfaceMode)
     Dispatch(daynaport, SET_IFACE_MODE);
 
     controller->SetCdbByte(5, DaynaPort::CMD_SCSILINK_STATS);
-    Dispatch(daynaport, SET_IFACE_MODE, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB,
-        "Not implemented");
+    Dispatch(daynaport, SET_IFACE_MODE, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB, "Not implemented");
 
     controller->SetCdbByte(5, DaynaPort::CMD_SCSILINK_ENABLE);
-    Dispatch(daynaport, SET_IFACE_MODE, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB,
-        "Not implemented");
+    Dispatch(daynaport, SET_IFACE_MODE, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB, "Not implemented");
 
     controller->SetCdbByte(5, DaynaPort::CMD_SCSILINK_SET);
-    Dispatch(daynaport, SET_IFACE_MODE, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB,
-        "Not implemented");
+    Dispatch(daynaport, SET_IFACE_MODE, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB, "Not implemented");
 }
 
 TEST_F(DaynaportTest, SetMcastAddr)
 {
-    Dispatch(daynaport, SET_MCAST_ADDR, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB,
-        "Length of 0 is not supported");
+    Dispatch(daynaport, SET_MCAST_ADDR, ILLEGAL_REQUEST, INVALID_FIELD_IN_CDB, "Length of 0 is not supported");
 
     controller->SetCdbByte(4, 1);
     EXPECT_CALL(*controller, DataOut);
